+++ runway/templates/cdk-tsc/bin/sample.ts
@@ -1,0 +1,7 @@
+#!/usr/bin/env node
+import 'source-map-support/register';
+import cdk = require('@aws-cdk/cdk');
+import { SampleStack } from '../lib/sample-stack';
+
+const app = new cdk.App();
+new SampleStack(app, 'SampleStack');