+++ runway/templates/cdk-tsc/lib/sample-stack.ts
@@ -1,0 +1,9 @@
+import cdk = require('@aws-cdk/cdk');
+
+export class SampleStack extends cdk.Stack {
+  constructor(scope: cdk.Construct, id: string, props?: cdk.StackProps) {
+    super(scope, id, props);
+
+    // The code that defines your stack goes here
+  }
+}